Задание 1 - Эллипс

+++ homework1/task1/ellipse.h
@@ -3,16 +3,16 @@
 class Ellipse
 {
     public:
-        double h, k, a, b;
+        int h, k, a, b;
 
     public:
-    Ellipse(double h = 0, double k = 0, double a = 0, double b = 0);
+    Ellipse(int h = 0, int k = 0, int a = 0, int b = 0);
     Ellipse(const Ellipse &e);
     ~Ellipse();
-    void normalize_axis(double &a, double &b);
+    void normalize_axis(int &a, int &b);
     double count_hyperparametr();
-    void set_all(double h_coord, double k_coord, double par_a, double par_b);
-    void get_all(double &h_coord, double &k_coord, double &par_a, double &par_b);
+    void set_all(int h_coord, int k_coord, int par_a, int par_b);
+    void get_all(int &h_coord, int &k_coord, int &par_a, int &par_b);
     void point_coordinates();
     void focus_coordinates();
     void count_loc_hord();
